Food route prints, but robot doesn't follow
Updated author

+++ CS301_Class.cyprj.Archive01-20190825T234813Z-001/CS301_Class.cyprj.Archive01/CS301_Class.cydsn/BFS/point.h
@@ -1,9 +1,11 @@
 #ifndef POINT_H
 #define POINT_H
+    
+#include <stdint.h>
 
 typedef struct point{
-    int x;
-    int y;
+    int8_t x;
+    int8_t y;
 } point;
 
 #endif
